Revert "moving changes (#960)"
This reverts commit 70d8a8036c52ae2f91aee081fe0d1c4c5451a58f.
