+++ genesyscloud/data_source_genesyscloud_architect_user_prompt_test.go
@@ -1,4 +1,4 @@
-package architect_user_prompt
+package genesyscloud
 
 import (
 	"fmt"
