+++ genesyscloud/resource_genesyscloud_init.go
@@ -18,6 +18,7 @@ func registerDataSources(l registrar.Registrar) {
 	l.RegisterDataSource("genesyscloud_location", DataSourceLocation())
 	l.RegisterDataSource("genesyscloud_auth_division_home", DataSourceAuthDivisionHome())
 	l.RegisterDataSource("genesyscloud_architect_schedules", DataSourceSchedule())
+	l.RegisterDataSource("genesyscloud_architect_user_prompt", dataSourceUserPrompt())
 	l.RegisterDataSource("genesyscloud_auth_division", dataSourceAuthDivision())
 	l.RegisterDataSource("genesyscloud_auth_division_home", DataSourceAuthDivisionHome())
 	l.RegisterDataSource("genesyscloud_journey_action_map", dataSourceJourneyActionMap())
@@ -48,6 +49,7 @@ func registerResources(l registrar.Registrar) {
 	l.RegisterResource("genesyscloud_routing_queue", ResourceRoutingQueue())
 	l.RegisterResource("genesyscloud_location", ResourceLocation())
 	l.RegisterResource("genesyscloud_architect_schedules", ResourceArchitectSchedules())
+	l.RegisterResource("genesyscloud_architect_user_prompt", ResourceArchitectUserPrompt())
 	l.RegisterResource("genesyscloud_auth_division", ResourceAuthDivision())
 	l.RegisterResource("genesyscloud_idp_adfs", ResourceIdpAdfs())
 	l.RegisterResource("genesyscloud_idp_generic", ResourceIdpGeneric())
@@ -87,6 +89,7 @@ func registerResources(l registrar.Registrar) {
 
 func registerExporters(l registrar.Registrar) {
 	l.RegisterExporter("genesyscloud_architect_schedules", ArchitectSchedulesExporter())
+	l.RegisterExporter("genesyscloud_architect_user_prompt", ArchitectUserPromptExporter())
 	l.RegisterExporter("genesyscloud_auth_division", AuthDivisionExporter())
 	l.RegisterExporter("genesyscloud_idp_adfs", IdpAdfsExporter())
 	l.RegisterExporter("genesyscloud_idp_generic", IdpGenericExporter())
