+++ genesyscloud/tfexporter/resource_genesyscloud_tf_export_test.go
@@ -28,8 +28,6 @@ import (
 
 	"github.com/google/uuid"
 
-	userPrompt "terraform-provider-genesyscloud/genesyscloud/architect_user_prompt"
-
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 	"gonum.org/v1/gonum/graph/simple"
@@ -991,7 +989,7 @@ func TestAccResourceTfExportUserPromptExportAudioFile(t *testing.T) {
 		exportTestDir    = "../.terraform" + uuid.NewString()
 	)
 
-	userPromptAsset := userPrompt.UserPromptResourceStruct{
+	userPromptAsset := gcloud.UserPromptResourceStruct{
 		Language:        userPromptResourceLanguage,
 		Tts_string:      util.NullValue,
 		Text:            strconv.Quote(userPromptResourceText),
@@ -999,7 +997,7 @@ func TestAccResourceTfExportUserPromptExportAudioFile(t *testing.T) {
 		FileContentHash: userResourcePromptFilename1,
 	}
 
-	userPromptAsset2 := userPrompt.UserPromptResourceStruct{
+	userPromptAsset2 := gcloud.UserPromptResourceStruct{
 		Language:        userPromptResourceLanguage2,
 		Tts_string:      util.NullValue,
 		Text:            strconv.Quote(userPromptResourceText2),
@@ -1007,8 +1005,8 @@ func TestAccResourceTfExportUserPromptExportAudioFile(t *testing.T) {
 		FileContentHash: userResourcePromptFilename2,
 	}
 
-	userPromptResources := []*userPrompt.UserPromptResourceStruct{&userPromptAsset}
-	userPromptResources2 := []*userPrompt.UserPromptResourceStruct{&userPromptAsset, &userPromptAsset2}
+	userPromptResources := []*gcloud.UserPromptResourceStruct{&userPromptAsset}
+	userPromptResources2 := []*gcloud.UserPromptResourceStruct{&userPromptAsset, &userPromptAsset2}
 
 	defer os.RemoveAll(exportTestDir)
 
@@ -1017,7 +1015,7 @@ func TestAccResourceTfExportUserPromptExportAudioFile(t *testing.T) {
 		ProviderFactories: provider.GetProviderFactories(providerResources, providerDataSources),
 		Steps: []resource.TestStep{
 			{
-				Config: userPrompt.GenerateUserPromptResource(&userPrompt.UserPromptStruct{
+				Config: gcloud.GenerateUserPromptResource(&gcloud.UserPromptStruct{
 					ResourceID:  userPromptResourceId,
 					Name:        userPromptName,
 					Description: strconv.Quote(userPromptDescription),
@@ -1036,7 +1034,7 @@ func TestAccResourceTfExportUserPromptExportAudioFile(t *testing.T) {
 					[]string{
 						strconv.Quote("genesyscloud_architect_user_prompt." + userPromptResourceId),
 					},
-				) + userPrompt.GenerateUserPromptResource(&userPrompt.UserPromptStruct{
+				) + gcloud.GenerateUserPromptResource(&gcloud.UserPromptStruct{
 					ResourceID:  userPromptResourceId,
 					Name:        userPromptName,
 					Description: strconv.Quote(userPromptDescription),
@@ -1053,7 +1051,7 @@ func TestAccResourceTfExportUserPromptExportAudioFile(t *testing.T) {
 			},
 			// Update to two resources with separate audio files
 			{
-				Config: userPrompt.GenerateUserPromptResource(&userPrompt.UserPromptStruct{
+				Config: gcloud.GenerateUserPromptResource(&gcloud.UserPromptStruct{
 					ResourceID:  userPromptResourceId,
 					Name:        userPromptName,
 					Description: strconv.Quote(userPromptDescription),
@@ -1072,7 +1070,7 @@ func TestAccResourceTfExportUserPromptExportAudioFile(t *testing.T) {
 					[]string{
 						strconv.Quote("genesyscloud_architect_user_prompt." + userPromptResourceId),
 					},
-				) + userPrompt.GenerateUserPromptResource(&userPrompt.UserPromptStruct{
+				) + gcloud.GenerateUserPromptResource(&gcloud.UserPromptStruct{
 					ResourceID:  userPromptResourceId,
 					Name:        userPromptName,
 					Description: strconv.Quote(userPromptDescription),
